+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Transfer speed (throughput) and elapsed time display in QR code receiver view
+
 ## [0.1.3] - 2026-03-04
 
 ### Fixed